Fix possible NPE seen in log from TISTUD-5756 NPE when closing a project

+++ plugins/com.aptana.editor.common/src/com/aptana/editor/common/AbstractThemeableEditor.java
@@ -613,8 +613,14 @@ public abstract class AbstractThemeableEditor extends AbstractFoldingEditor impl
 			if (dndService != null)
 			{
 				ProjectionViewer viewer = (ProjectionViewer) getSourceViewer();
-				StyledText st = viewer.getTextWidget();
-				dndService.removeMergedDropTarget(st);
+				if (viewer != null)
+				{
+					StyledText st = viewer.getTextWidget();
+					if (st != null)
+					{
+						dndService.removeMergedDropTarget(st);
+					}
+				}
 			}
 		}
 		finally
